+++ Zapt/Categories/NSArray+Functional.m
@@ -284,7 +284,7 @@
 
 - (id)randomObject
 {
-    NSUInteger idx = arc4random_uniform(self.count);
+    NSUInteger idx = arc4random_uniform((unsigned int)self.count);
     
     return self[idx];
 }
