+++ Zapt/View/FAUIElementWithActivity.h
@@ -16,7 +16,7 @@
 
 @optional
 @property (atomic, retain, readwrite) UIActivityIndicatorView *activityIndicatorView;
-@property int startCount;
+@property NSInteger startCount;
 - (void)startActivityWithCount:(NSInteger)count; // increments startcount by count
 - (void)stopAllActivity; // Stops all activity no matter what the count was before
